fix(ui): satisfy FilterFns type declaration in StakingTable
