+++ ui/src/interfaces/table.d.ts
@@ -1,0 +1,7 @@
+import { FilterFn } from '@tanstack/react-table'
+
+declare module '@tanstack/react-table' {
+  interface FilterFns {
+    global: FilterFn<unknown>
+  }
+}
